iris/requirements.txt
black==24.2.0
fastapi==0.110.0
flake8==7.0.0
langchain==0.1.11
openai==1.13.3
pre-commit==3.6.2
pydantic==2.6.3
PyYAML==6.0.1
uvicorn==0.27.1
requests~=2.31.0
weaviate-client==4.5.4
PyMuPDF=1.23.22